# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CareerPath AI Architect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
...
  /work-experiences:
    get:
      operationId: listWorkExperiences
      tags: [work-experience]
      summary: List work experiences
      responses:
        "200":
          description: List of work experiences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkExperience"
    post:
      operationId: createWorkExperience
      tags: [work-experience]
      summary: Create a work experience entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkExperienceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkExperience"
  /work-experiences/{id}:
    patch:
      operationId: updateWorkExperience
      tags: [work-experience]
      summary: Update a work experience entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkExperienceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkExperience"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteWorkExperience
      tags: [work-experience]
      summary: Delete a work experience entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List user skills
      responses:
        "200":
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [skills]
      summary: Add a skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSkillBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /skills/{id}:
    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Remove a skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /milestones/{id}/complete:
    patch:
      operationId: completeMilestone
      tags: [milestones]
      summary: Mark milestone as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        currentRole:
          type: ["string", "null"]
        totalExperienceMonths:
          type: ["integer", "null"]
        industry:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        professionalSummary:
          type: ["string", "null"]
        preferredLearningStyle:
          type: ["string", "null"]
        weeklyLearningMinutes:
          type: ["integer", "null"]
        salaryAspiration:
          type: ["string", "null"]
        careerLevel:
          type: ["string", "null"]
        cvImportCompletedAt:
          type: ["string", "null"]
        updatedAt:
          type: string
      required: [id, userId, updatedAt]
    UpdateProfileBody:
      type: object
      properties:
        currentRole:
          type: string
        totalExperienceMonths:
          type: integer
        industry:
          type: string
        location:
          type: string
        phone:
          type: string
        linkedinUrl:
          type: string
        professionalSummary:
          type: string
        preferredLearningStyle:
          type: string
        weeklyLearningMinutes:
          type: integer
        salaryAspiration:
          type: string
        careerLevel:
          type: string
        cvImportCompletedAt:
          type: string
...
    WorkExperience:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        company:
          type: string
        title:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        isCurrent:
          type: boolean
        description:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, company, title, startDate, isCurrent, createdAt]
    CreateWorkExperienceBody:
      type: object
      properties:
        company:
          type: string
        title:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isCurrent:
          type: boolean
        description:
          type: string
        skills:
          type: string
      required: [company, title, startDate, isCurrent]
...
    Skill:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        category:
          type: string
        proficiencyLevel:
          type: string
        createdAt:
          type: string
      required: [id, userId, name, category, proficiencyLevel, createdAt]
    CreateSkillBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        proficiencyLevel:
          type: string
      required: [name, category, proficiencyLevel]
...
    Milestone:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        analysisId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        phase:
          type: string
        completed:
          type: boolean
        dueDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, title, phase, completed, createdAt]
...